+++ plugins/mongodb/superduper_mongodb/__init__.py
@@ -4,7 +4,7 @@ from .metadata import MongoMetaDataStore as MetaDataStore
 from .query import MongoQuery
 from .vector_search import MongoAtlasVectorSearcher as VectorSearcher
 
-__version__ = "0.4.0"
+__version__ = "0.4.1"
 
 __all__ = [
     "ArtifactStore",